requirements.txt
ConfigParser
python-statsd
